package/libzenoh-pico: disable on or1k
The upstream libzenoh-pico code uses asm("nop"), which is completely
bogus, and actually breaks build on or1k as the instruction does not
exist.

In upstream commit
https://github.com/eclipse-zenoh/zenoh-pico/commit/7d79a14addcea972e801709413f5c9a5226e3ad3
they made the asm("nop") thing "configurable" with ZP_ASM_NOP, but
still that requires some special handling for or1k. The whole thing is
a complete mess, and upstream should have to use any of that. Rather
than stacking piles of bogus stuff on more bogus stuff, let's simply
disable this package for or1k for now.

Thanks a lot to Fabrice for all the investigation, and interaction
with upstream. See
https://patchwork.ozlabs.org/project/buildroot/patch/20240110215515.1263859-1-fontaine.fabrice@gmail.com/
for example.

Fixes:
 - http://autobuild.buildroot.org/results/fd0b2c666a1dc1537162d15b27743abd270243ed

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>