package/davinci-bootcount: rename to uboot-bootcount
Born as a project to read/write the bootcount on the TI am3xx platform,
over time it has introduced bootcount management on EEPROM and for the
stm32mp1 platform. As a result, the project removed the 'davinci' tag
from its name and GitHub link.
The patch aligns the package name in Buildroot with the current one on
GitHub.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
[Peter: extend/rework legacy handling]
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>