package/jailhouse: install files to /usr
The jailhouse package installs files in the /usr/local prefix. This is
not convenient, because the jailhouse binary ends up in /usr/local/sbin,
which is not in the default Buildroot PATH. See [1]. Moreover, all
other Buildroot packages install files in /usr.

This is because the Buildroot package recipe does not set any prefix
value, and the upstream Jailhouse Makefile has a default to /usr/local.
See [2].

This commit sets the prefix value in the JAILHOUSE_MAKE_OPTS and also
updates other installation commands to put all the files in /usr.

[1] https://gitlab.com/buildroot.org/buildroot/-/blob/2023.11.1/system/Config.in#L495
[2] https://github.com/siemens/jailhouse/blob/v0.12/scripts/include.mk#L22

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>