Revert "docs/website/support.html: document irc channel permission changes"
This didn't work out as planned, neither the restriction of muting
unregistered users, nor the exception for matrix users worked as planned.

The channel mode has been reverted to +R (meaning only registered users
are allowed to join) and an exception for *that* has been introduced for
matrix users via +e. The channel modes are documented in [1].

[1] https://www.oftc.net/ChannelModes/

This reverts commit d1e6d7845b0f3b450709b84a6e4fc4527f12b196.

Signed-off-by: Marcus Hoffmann <buildroot@bubu1.eu>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>