package/atop: bump to version 2.10.0
This new version requires libglib2 as a dependency, which itself is
detected using pkg-config, so we need host-pkgconf as a dependency.

The patch 0001-atop.h-include-time.h.patch can be removed, it was
applied upstream as
https://github.com/Atoptool/atop/commit/03c5c4dd29bc7590baccb55a36d61e7f4b39c1fe.

A new patch (submitted upstream) is also needed to be able to provide
additional CFLAGS/LDFLAGS, without overriding the ones defined in the
upstream Makefile.

Signed-off-by: Kadambini Nema <kadambini.nema@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>