package/python-terminaltables: use correct pep517 build backend
We need to migrate python-terminaltables to the pep517 poetry-core
backend as setuptools is not supported when building with a pep517
frontend.

This package currently builds using setuptools as we do not yet
use setuptools with a pep517 build frontend. The package contains
a setuptools fallback which only can be used when using setuptools
without a pep517 frontend as the pep517 frontend will only use the
build backend specified in the package pyproject.toml which is
poetry-core and not setuptools. Thus, specifying setuptools as the build
backend is simply wrong.

The current release of python-terminaltables still uses poetry rather
than poetry-core as a build backend. poetry is much more heavyweight, it
would need to pull in a large number of build dependencies. Therefore,
include an upstream patch to switch from poetry to poetry-core.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
[Arnout: use a patch instead of sed]
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>