package/rust/rust.mk: add missing host-zlib dependency
Several rust tools are linking against zlib, so add the depedency
explicitly in HOST_RUST_DEPENDENCIES.

For now, host-rust build system is not able to find zlib provided
by Buildroot in HOST_DIR due to at least two issues that will be
fixed in followup commits.

Note that host-zlib is already in the dependency chain, by way of
host-openssl, but since rust needs for itself, we need to add it
as an explicit dependency.

Signed-off-by: Romain Naour <romain.naour@smile.fr>
[yann.morin.1998@free.fr: add not about transitive dependency]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>